pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "retsu"
version = "0.2.4"  # semantic-release
description = "Retsu aims to wrap-up Celery in way to facilitate to create parallel and serial tasks"
readme = "README.md"
authors = ["Ivan Ogasawara <ivan.ogasawara@gmail.com>"]
packages = [
  {include = "retsu", from="src"},
]
license = "BSD 3 Clause"
exclude = [
  ".git/*",
  ".env*",
]
include = ["src/retsu/py.typed"]


[tool.poetry.dependencies]
python = ">=3.8.1,<4"
atpublic = ">=4.0"
celery = ">=5"
redis = ">=5"
django = { version = ">=3", optional = true }
typing-extensions = ">=4.12.0"
asyncio = ">=3.4.3"

[tool.poetry.extras]
django = [
  "django",
]


[tool.poetry.group.dev.dependencies]
pytest = ">=7.3.2"
pytest-cov = ">=4.1.0"
coverage = ">=7.2.7"
pre-commit = ">=3.3.2"
ruff = ">=0.2.0"
mypy = ">=1.5.0"
bandit = ">=1.7.5"
vulture = ">=2.7"
mccabe = ">=0.6.1"
ipython = "<8"
ipykernel = ">=6.0.0"
Jinja2 = ">=3.1.2"
mkdocs = ">=1.4.3"
mkdocs-exclude = ">=1.0.2"
mkdocs-jupyter = ">=0.24.1"
mkdocs-literate-nav = ">=0.6.0"
mkdocs-macros-plugin = ">=0.7.0,<1"
mkdocs-material = ">=9.1.15"
mkdocstrings = ">=0.21.2"
mkdocstrings-python = ">=1.1.2"
mkdocs-gen-files = ">=0.5.0"
makim = "1.15.1"
# 'PosixPath' object has no attribute 'endswith'
virtualenv = "<=20.25.1"
flask = ">=3"
containers-sugar = "1.13.0"
compose-go = "2.27.0"
django = ">=3"
django-stubs = ">=3"
pytest-asyncio = ">=0.23.8"

[tool.pytest.ini_options]
# asyncio_mode = "auto"
testpaths = [
    "tests",
]

[tool.bandit]
exclude_dirs = ["tests"]
targets = "./"

[tool.vulture]
exclude = ["tests"]
ignore_decorators = []
ignore_names = []
make_whitelist = true
min_confidence = 80
paths = ["./"]
sort_by_size = true
verbose = false

[tool.ruff]
line-length = 79
force-exclude = true
src = ["./"]
exclude = [
  'docs',
]
fix = true

[tool.ruff.lint]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "D",   # pydocstyle
  "YTT", # flake8-2020
  "RUF", # Ruff-specific rules
  "I001",  # isort
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.isort]
# Use a single line between direct and from import
lines-between-types = 1

[tool.mypy]
python_version = "3.8"
check_untyped_defs = true
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
exclude = ["examples/", "scripts/", "src/retsu/asyncio/"]
